Remove panelizer choice from unused view_modes.
This fixes the problem whereby upon save basic page nodes would
not preserve any customizations made to the panelizer display.

This seems to be a problem with features in that these settings
were retained in the feature even after the default and teaser
view modes were set to not be panelized.